fix(code-gen/java): Updated the codegen for java to generally use smaller types for unsigned integers (a byte is able to keep an unsigned integer value to 7 bits and not up to 4 bits)

+++ plc4j/drivers/profinet-ng/src/main/java/org/apache/plc4x/java/profinet/context/ProfinetDriverContext.java
@@ -264,7 +264,7 @@ public class ProfinetDriverContext implements DriverContext {
     }
 
     public DceRpc_ObjectUuid getCmInitiatorObjectUuid() {
-        return new DceRpc_ObjectUuid((byte) 0x00, 0x0001,
+        return new DceRpc_ObjectUuid((byte) 0x00, (short) 0x0001,
             Integer.decode("0x" + getDeviceId()),
             Integer.decode("0x" + getVendorId()));
     }
